Fix PostgreSQL CI errors (#706)
* do not ignore PostgreSQL CI errors

* "like with variable" is not supported in PostgresSQL

* fix field type in test

* fix wrongly defined type for Migrator

* fix unsupported SUM(LENGTH("name")) and SUM(5)

* fix type needed for SUM

* call lastInsertId before HOOK_AFTER_INSERT_QUERY

* do not call lastInsertId() if ID explicitly set

* synchronize ID sequence for PostgreSQL after insert with explicit ID

* fix back and forth call

* sync ID also before insert

* insert returns only string ID

* fix random order

* proper expr escape

* keep only really needed tests to be skipped on PostgreSQL

* fix typo in doc

+++ tests/DeepCopyTest.php
@@ -88,7 +88,7 @@ class DcInvoiceLine extends Model
 
         $this->addField('qty', ['type' => 'integer', 'mandatory' => true]);
         $this->addField('price', ['type' => 'money']);
-        $this->addField('vat', ['type' => 'numeric', 'default' => 0.21]);
+        $this->addField('vat', ['type' => 'float', 'default' => 0.21]);
 
         // total is calculated with VAT
         $this->addExpression('total', '[qty]*[price]*(1+[vat])');
